requirements.txt
monday>=2.0.0
PyQt6>=6.4.0
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.28.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
cryptography>=41.0.0
pillow>=10.0.0 